Add TypeScript support to Atom (again)
Ok, so it turns out my version of atom was very outdated, and did not
support the atom-ide-ui package, so I had to rely on the other bunch of
packages to support developing TypeScript applications, which I wasn't
very happy with.

Now my version of atom is updated and does support the package, so I'm
adding it now.